Added in collision behavior for root
I added a "StageBoundaries" prefab, which has a line renderer. Just set
the line renderer to the bounds of the stage, and the code takes care of
the rest.

+++ Dig Deep Sketch/Assets/Scripts/RootSystem.cs
@@ -66,7 +66,7 @@ public class RootSystem : MonoBehaviour {
                     if (branch && _selectedPoint != _selectedBranchPosition)
                     {
                         Vector3[] oldRootPath = _roots[i].GetLRPositions();
-                        int _lastOldRootAngle = _roots[i]._lastAngle;
+                        float _lastOldRootAngle = _roots[i]._lastAngle;
                         RootPath firstRootBranch = GameObject.Instantiate(Root);
                         RootPath secondRootBranch = GameObject.Instantiate(Root);
                         firstRootBranch.SetRootPoint(Vector3.zero);
